requirements.txt
# FastAPI & Web Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# LangChain & OpenAI
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.24
openai==1.12.0
tiktoken==0.5.2

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# LangGraph
langgraph==0.0.26

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
docx2txt==0.8
unstructured==0.12.4

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Data & ML
pydantic==2.6.1
pydantic-settings==2.1.0
scikit-learn==1.4.0
numpy==1.26.4

# Policy & Config
pyyaml==6.0.1
python-dotenv==1.0.1

# Utilities
httpx==0.26.0
redis==5.0.1
tenacity==8.2.3
requests==2.31.0

# Testing & Dev
pytest==8.0.0
pytest-asyncio
pytest-cov==4.1.0
mypy==1.8.0
black==24.1.1
ruff==0.2.1

# Monitoring
prometheus-client==0.19.0
